Move link type setting to checkout session
