+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/checkout/settings/CheckoutPlaygroundSettings.kt
@@ -106,7 +106,7 @@ internal class CheckoutPlaygroundSettings private constructor(
         }
 
         fun applyFeatureFlags() {
-            this[CheckoutPlaygroundDefinitions.Controller.linkType].applyFeatureFlags()
+            this[CheckoutSessionDefinitions.linkType].applyFeatureFlags()
         }
     }
